+++ examples/screws.py
@@ -1,7 +1,7 @@
 from string import digits
 from typing import Optional
 
-from labelfrontend.units import LengthDimension, mm
+from pysvglabel.labelfrontend import LengthDimension, mm
 
 
 # Abbreviations from
