+++ pysvglabel/labelcore/__init__.py
