+++ pysvglabel/labelcore/common.py
