+++ pysvglabel/labelfrontend/Code128.py
@@ -1,12 +1,11 @@
-from typing import List, cast
+from typing import List
 import xml.etree.ElementTree as ET
 
 from .Align import Align
 from .external.Code128 import code128_widths
 
-from labelcore.SvgTemplate import SvgTemplate
-from labelcore.common import SVG_NAMESPACE
-from labelcore.GroupReplacer import RectGroupReplacer
+from pysvglabel.labelcore.common import SVG_NAMESPACE
+from pysvglabel.labelcore.GroupReplacer import RectGroupReplacer
 from .units import LengthDimension, px
