+++ pysvglabel/labelfrontend/Scaling.py
