+++ pysvglabel/printer/__main__.py
@@ -8,7 +8,7 @@ from typing import Dict, Tuple
 import win32com.shell.shell as shell  # type:ignore
 import win32event  # type:ignore
 
-from labelcore import SvgTemplate, InkscapeSubprocess
+from ..labelcore import SvgTemplate, InkscapeSubprocess
 
 
 if __name__ == '__main__':
