--- requirements.txt
@@ -1,2 +1,0 @@
-ppf-datamatrix==0.2
-qrcode==7.4.2
