+++ tests/test_simple.py
@@ -1,8 +1,8 @@
 import csv
 
 import os.path
-from labelcore import SvgTemplate, NAMESPACES
-from labelcore.SvgTemplate import get_text_of
+from pysvglabel.labelcore import SvgTemplate, NAMESPACES
+from pysvglabel.labelcore.SvgTemplate import get_text_of
 from .LabelTestCase import LabelTestCase
